Move the congestion control interface into mod.rs
