+++ neqo-transport/src/cc/mod.rs
@@ -7,12 +7,51 @@
 // Congestion control
 #![deny(clippy::pedantic)]
 
-pub mod cc;
+use crate::tracking::SentPacket;
+use neqo_common::qlog::NeqoQlog;
+
+use std::fmt::Debug;
+use std::fmt::Display;
+use std::time::{Duration, Instant};
+
 mod new_reno_cc;
 
-pub use cc::CongestionControl;
 use new_reno_cc::NewReno;
 pub use new_reno_cc::{CWND_INITIAL_PKTS, CWND_MIN, MAX_DATAGRAM_SIZE, PACING_BURST_SIZE};
+
+pub trait CongestionControl: Display + Debug {
+    fn set_qlog(&mut self, qlog: NeqoQlog);
+
+    #[cfg(test)]
+    fn cwnd(&self) -> usize;
+
+    #[cfg(test)]
+    fn ssthresh(&self) -> usize;
+
+    #[cfg(test)]
+    fn bif(&self) -> usize;
+
+    fn cwnd_avail(&self) -> usize;
+
+    fn on_packets_acked(&mut self, acked_pkts: &[SentPacket]);
+
+    fn on_packets_lost(
+        &mut self,
+        now: Instant,
+        first_rtt_sample_time: Option<Instant>,
+        prev_largest_acked_sent: Option<Instant>,
+        pto: Duration,
+        lost_packets: &[SentPacket],
+    );
+
+    fn discard(&mut self, pkt: &SentPacket);
+
+    fn on_packet_sent(&mut self, pkt: &SentPacket, rtt: Duration);
+
+    fn start_pacer(&mut self, now: Instant);
+
+    fn next_paced(&self, rtt: Duration) -> Option<Instant>;
+}
 
 pub enum CongestionControlAlgorithm {
     NewReno,
